requirements.txt
opencv-python
numpy
tensorflow
pillow
streamlit
matplotlib
scikit-learn
pytesseract
python-Levenshtein
deep-translator
